fix(masonry): rename renderWidget to widget + move it to a seprate file within the same folder
